requirements.txt
asgiref==3.9.1
dj-database-url==2.3.0
Django==5.2.6
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
gunicorn==21.2.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.11.0
